feat: refactored dashboard ir out of pages and in app
