+++ src/app/Dashboard/Drawer/Drawer.jsx
@@ -9,7 +9,7 @@ import List from "@material-ui/core/List";
 import Divider from "@material-ui/core/Divider";
 import IconButton from "@material-ui/core/IconButton";
 
-import { Button } from "../../../component";
+import { Button } from "../../component";
 import { Grid } from "@material-ui/core";
 import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
 import ListItem from "@material-ui/core/ListItem";
