+++ src/app/Dashboard/Home/Overview.jsx
@@ -3,8 +3,8 @@ import { makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
 import Card from "@material-ui/core/Card";
-import { Mission } from "../../../model";
-import { H2, H4 } from "../../../component";
+import { Mission } from "../../model";
+import { H2, H4 } from "../../component";
 import { compose } from "redux";
 import { connect } from "react-redux";
 import { firestoreConnect } from "react-redux-firebase";
